+++ apps/meteor/ee/app/livechat-enterprise/server/lib/units.ts
@@ -8,7 +8,7 @@ async function hasUnits(): Promise<boolean> {
 }
 
 // Units should't change really often, so we can cache the result
-const memoizedHasUnits = mem(hasUnits, { maxAge: 10000 });
+const memoizedHasUnits = mem(hasUnits, { maxAge: process.env.TEST_MODE ? 1 : 10000 });
 
 export async function getUnitsFromUser(): Promise<{ [k: string]: any }[] | undefined> {
 	if (!(await memoizedHasUnits())) {