+++ ee/packages/pdf-worker/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @rocket.chat/pdf-worker
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [239a34e877]
+- Updated dependencies [203304782f]
+- Updated dependencies [4186eecf05]
+- Updated dependencies [ba24f3c21f]
+- Updated dependencies [ebab8c4dd8]
+- Updated dependencies [1041d4d361]
+- Updated dependencies [61128364d6]
+- Updated dependencies [d45365436e]
+  - @rocket.chat/core-typings@6.4.0
+
+## 0.0.15-rc.5
+
+### Patch Changes
+
+- Updated dependencies [1041d4d361]
+  - @rocket.chat/core-typings@6.4.0-rc.5
+
 ## 0.0.14-rc.4
 
 ### Patch Changes