+++ packages/instance-status/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rocket.chat/instance-status
 
+## 0.0.15
+
+### Patch Changes
+
+- @rocket.chat/models@0.0.15
+
+## 0.0.15-rc.5
+
+### Patch Changes
+
+- @rocket.chat/models@0.0.15-rc.5
+
 ## 0.0.14-rc.4
 
 ### Patch Changes